Refactored 'misc_utils' into 'helpers.py'
